requirements.txt
django==1.11.17
djangorestframework==3.10.3
django-ckeditor==5.4.0
django-cleanup==2.1.0
django-crispy-forms==1.7.2
django-js-asset==1.0.0
dj-database-url==0.4.0
dj-static==0.0.6
google-cloud-vision==2.1.0
gunicorn==19.4.5
numpy==1.20.0
opencv-python==4.1.2.30
whitenoise==2.0.6